refactor doughnut and announcement
