+++ assembl/static2/css/_assembl_base_styles.scss
@@ -2,6 +2,7 @@
 
 //@import 'variables'; // The current theme imports the default theme which imports variables.scss. The current theme then overrides some of these variables.
 @import 'variables_cascade';
+@import 'components/announcement';
 @import 'components/alert';
 @import 'components/assembl-glyph';
 @import 'components/avatar';
@@ -19,6 +20,7 @@
 @import 'components/scrollOnePageButton';
 @import 'components/sentiments';
 @import 'components/socialMedia';
+@import 'components/statistics-doughnut';
 @import 'components/text';
 @import 'components/thumbnail';
 @import 'components/timeline';
