+++ assembl/static2/css/components/announcement.scss
@@ -1,57 +1,3 @@
-@import '../variables.scss';
-
-.no-arrow-tooltip {
-  .tooltip-arrow {
-    display: none;
-  }
-}
-
-.statistics-container{
-  text-align: center;
-  display: inline-block;
-  width: 100%;
-  height: 100%;
-
-  .statistics {
-    position: relative;
-    width: 100%;
-    height: 100%;
-
-    svg {
-      width: 100%;
-      height: 100%;
-    }
-
-    .circle {
-      stroke-width: 16px;
-    }
-
-    .superpose {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-    }
-
-    .superpose-label {
-      top: calc(50% - 1.5em);
-
-      .doughnut-label-count {
-        font-family: $font-family3;
-        font-size: $font-size-sm;
-        letter-spacing: initial;
-      }
-
-      .doughnut-label-text {
-        font-family: $font-family4;
-        font-size: $font-size-xxs;
-        letter-spacing: initial;
-      }
-    }
-  }
-}
-
 .announcement {
   text-align: right;
 
@@ -63,11 +9,13 @@
     text-align: left;
   }
 
-  .announcement-left {
+  .announcement-statistics {
+    display: inline-block;
+    text-align: center;
 
-    .announcement-bottom {
-      display: inline-block;
-      text-align: center;
+    .announcement-doughnut {
+      width: 150px;
+      height: 150px;
 
       .superpose-label {
         top: calc(50% - 1.7em);
@@ -82,30 +30,34 @@
           letter-spacing: 0.08em;
         }
       }
+    }
 
-      .announcement-numbers {
-        font-size: 12px;
-        margin-bottom: 15%;
+    .announcement-numbers {
+      font-size: 12px;
+      margin-bottom: 15%;
+      color: $first-color;
 
-        span::before {
-          font-size: 14px;
-        }
+      span::before {
+        font-size: 14px;
       }
     }
   }
 
-  .announcement-right {
-    .col-sm-8 {
-      margin-left: initial;
-    }
-  }
+  .announcement-video {
 
-  .video-section {
-    div {
-      background-color: $grey0;
-    }
-    .video-title-section {
-      display: none;
+    .video-section {
+
+      .col-sm-8 {
+        margin-left: initial;
+      }
+
+      div {
+        background-color: $grey0;
+      }
+
+      .video-title-section {
+        display: none;
+      }
     }
   }
 }
