+++ assembl/static2/css/components/statistics-doughnut.scss
@@ -1,0 +1,45 @@
+.statistics-doughnut {
+  text-align: center;
+  display: inline-block;
+  width: 100%;
+  height: 100%;
+
+  .statistics {
+    position: relative;
+    width: 100%;
+    height: 100%;
+
+    svg {
+      width: 100%;
+      height: 100%;
+    }
+
+    .circle {
+      stroke-width: 16px;
+    }
+
+    .superpose {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+    }
+
+    .superpose-label {
+      top: calc(50% - 1.5em);
+
+      .doughnut-label-count {
+        font-family: $font-family3;
+        font-size: $font-size-sm;
+        letter-spacing: initial;
+      }
+
+      .doughnut-label-text {
+        font-family: $font-family4;
+        font-size: $font-size-xxs;
+        letter-spacing: initial;
+      }
+    }
+  }
+}
